Adding login controller
